pyproject.toml
[tool.poetry]
name = "ktb-fund-manager"
version = "1.0.0"
description = "KTB Fund Manager - An AI-powered hedge fund for educational purposes"
authors = ["KTB <kboyle89@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.7"
langchain-anthropic = "^0.3.5"
langchain-openai = "^0.3.5"
langgraph = "^0.2.56"
pandas = "^2.1.0"
numpy = "^1.24.0"
python-dotenv = "^1.0.0"
matplotlib = "^3.9.2"
tabulate = "^0.9.0"
colorama = "^0.4.6"
questionary = "^2.1.0"
rich = "^13.9.4"
httpx = "^0.27.0"
pydantic = "^2.4.2"
fastapi = {extras = ["standard"], version = "^0.104.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
force_alphabetical_sort_within_sections = true
